docs(form): merge validation page
