+++ site/internal/pages/demo/layout.templ
@@ -487,7 +487,6 @@ func getSidebarSections(activeComponent string) []sidebar.Section {
 				sItem("dropdown", "Dropdown", "/components/dropdown", activeComponent),
 				sItem("fileinput", "File Input", "/components/fileinput", activeComponent),
 				sItem("form", "Form", "/components/form", activeComponent),
-				sItem("form-validation", "Form Validation", "/components/form-validation", activeComponent),
 				sItem("pagination", "Pagination", "/components/pagination", activeComponent),
 				sItem("radio", "Radio", "/components/radio", activeComponent),
 				sItem("palette", "Palette", "/components/palette", activeComponent),
@@ -556,7 +555,6 @@ var orderedComponents = []componentNavLink{
 	{"Dropdown", "/components/dropdown"},
 	{"File Input", "/components/fileinput"},
 	{"Form", "/components/form"},
-	{"Form Validation", "/components/form-validation"},
 	{"Pagination", "/components/pagination"},
 	{"Select", "/components/select"},
 	{"Palette", "/components/palette"},
